+++ build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java'
 }
 
-group = 'camp.nextstep.edu'
+group = 'com.kakao.onboarding.precourse.dukeyun'
 version = '1.0-SNAPSHOT'
 
 java {
