--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/App.java
@@ -1,12 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball;
-
-import com.kakao.onboarding.precource.dukeyun.baseball.framework.gamesystem.GameSystem;
-
-public class App {
-
-	public static void main(String[] args) {
-		GameSystem baseballGameSystem = GameFactory.createBaseballGameSystem();
-		baseballGameSystem.run();
-	}
-
-}
