--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/baseball/BaseballNumberGenerator.java
@@ -1,5 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball.baseball;
-
-public interface BaseballNumberGenerator {
-	BaseballNumber generate();
-}
