--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/framework/game/turn/TurnBaseGameService.java
@@ -1,9 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball.framework.game.turn;
-
-public interface TurnBaseGameService {
-
-	void init();
-
-	TurnResult playTurn(TurnInput turnInput);
-
-}
