--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/framework/game/turn/TurnBaseGameView.java
@@ -1,13 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball.framework.game.turn;
-
-public interface TurnBaseGameView {
-
-	void printStart();
-
-	TurnInput requestTurnInput();
-
-	void printClear();
-
-	void printResult(TurnResult turnResult);
-
-}
