--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/framework/game/turn/TurnInput.java
@@ -1,4 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball.framework.game.turn;
-
-public interface TurnInput {
-}
