--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/framework/game/turn/TurnResult.java
@@ -1,7 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball.framework.game.turn;
-
-public interface TurnResult {
-
-	boolean isGameCleared();
-
-}
