--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/framework/gamesystem/GameSystem.java
@@ -1,6 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball.framework.gamesystem;
-
-public interface GameSystem {
-
-	void run();
-}
