--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/framework/gamesystem/ReplayableGameSystemView.java
@@ -1,7 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball.framework.gamesystem;
-
-public interface ReplayableGameSystemView {
-
-	boolean requestReplayInput();
-
-}
