--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/utils/Console.java
@@ -1,12 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball.utils;
-
-import java.util.Scanner;
-
-public class Console {
-
-	private static final Scanner scanner = new Scanner(System.in);
-
-	public static String readLine() {
-		return scanner.nextLine().trim();
-	}
-}
