+++ src/main/java/com/kakao/onboarding/precourse/dukeyun/baseball/baseball/BaseballNumberGenerator.java
@@ -1,0 +1,5 @@
+package com.kakao.onboarding.precourse.dukeyun.baseball.baseball;
+
+public interface BaseballNumberGenerator {
+	BaseballNumber generate();
+}
