+++ src/main/java/com/kakao/onboarding/precourse/dukeyun/baseball/baseball/RandomBaseballNumberGenerator.java
@@ -1,0 +1,30 @@
+package com.kakao.onboarding.precourse.dukeyun.baseball.baseball;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Random;
+
+public class RandomBaseballNumberGenerator implements BaseballNumberGenerator {
+
+	private static final Random RANDOM = new Random();
+
+	private final BaseballGameConfig config;
+
+	public RandomBaseballNumberGenerator(BaseballGameConfig config) {
+		this.config = config;
+	}
+
+	@Override
+	public BaseballNumber generate() {
+		List<Integer> numbers = new ArrayList<>();
+		while (numbers.size() < config.getLength()) {
+			int randomNumber = RANDOM.nextInt(config.getMinRange(), config.getMaxRange() + 1);
+			if (!numbers.contains(randomNumber)) {
+				numbers.add(randomNumber);
+			}
+		}
+
+		return new BaseballNumber(config, numbers);
+	}
+
+}
