+++ src/main/java/com/kakao/onboarding/precourse/dukeyun/baseball/framework/game/turn/TurnBaseGameService.java
@@ -1,0 +1,9 @@
+package com.kakao.onboarding.precourse.dukeyun.baseball.framework.game.turn;
+
+public interface TurnBaseGameService {
+
+	void init();
+
+	TurnResult playTurn(TurnInput turnInput);
+
+}
