+++ src/main/java/com/kakao/onboarding/precourse/dukeyun/baseball/framework/game/turn/TurnBaseGameView.java
@@ -1,0 +1,13 @@
+package com.kakao.onboarding.precourse.dukeyun.baseball.framework.game.turn;
+
+public interface TurnBaseGameView {
+
+	void printStart();
+
+	TurnInput requestTurnInput();
+
+	void printClear();
+
+	void printResult(TurnResult turnResult);
+
+}
