+++ src/main/java/com/kakao/onboarding/precourse/dukeyun/baseball/framework/game/turn/TurnInput.java
@@ -1,0 +1,4 @@
+package com.kakao.onboarding.precourse.dukeyun.baseball.framework.game.turn;
+
+public interface TurnInput {
+}
