+++ src/main/java/com/kakao/onboarding/precourse/dukeyun/baseball/framework/game/turn/TurnResult.java
@@ -1,0 +1,7 @@
+package com.kakao.onboarding.precourse.dukeyun.baseball.framework.game.turn;
+
+public interface TurnResult {
+
+	boolean isGameCleared();
+
+}
