+++ src/main/java/com/kakao/onboarding/precourse/dukeyun/baseball/framework/gamesystem/GameSystem.java
@@ -1,0 +1,6 @@
+package com.kakao.onboarding.precourse.dukeyun.baseball.framework.gamesystem;
+
+public interface GameSystem {
+
+	void run();
+}
