+++ src/main/java/com/kakao/onboarding/precourse/dukeyun/baseball/framework/gamesystem/ReplayableGameSystemView.java
@@ -1,0 +1,7 @@
+package com.kakao.onboarding.precourse.dukeyun.baseball.framework.gamesystem;
+
+public interface ReplayableGameSystemView {
+
+	boolean requestReplayInput();
+
+}
