+++ src/main/java/com/kakao/onboarding/precourse/dukeyun/baseball/utils/Console.java
@@ -1,0 +1,12 @@
+package com.kakao.onboarding.precourse.dukeyun.baseball.utils;
+
+import java.util.Scanner;
+
+public class Console {
+
+	private static final Scanner scanner = new Scanner(System.in);
+
+	public static String readLine() {
+		return scanner.nextLine().trim();
+	}
+}
